requirements.txt
flask
langchain
faiss-cpu
openai
python-dotenv
tiktoken
